+++ includes/features/class-wpcli.php
@@ -20,9 +20,7 @@ use OPcacheManager\Plugin\Opcache_Manager_Admin;
 use Spyc;
 
 /**
- * WP-CLI for OPcache Manager.
- *
- * Defines methods and properties for WP-CLI commands.
+ * Manages OPcache and get analytics about its use.
  *
  * @package Features
  * @author  Pierre Lannoy <https://pierre.lannoy.fr/>.
@@ -36,7 +34,7 @@ class Wpcli {
 	 * @since    2.0.0
 	 * @var array $exit_codes Exit codes.
 	 */
-	private static $exit_codes = [
+	private $exit_codes = [
 		0   => 'operation successful.',
 		1   => 'unrecognized setting.',
 		2   => 'unrecognized action.',
@@ -51,7 +49,7 @@ class Wpcli {
 	 * @param   string  $field  Optional. The field to output.
 	 * @since   2.0.0
 	 */
-	private static function write_ids( $ids, $field = '' ) {
+	private function write_ids( $ids, $field = '' ) {
 		$result = '';
 		$last   = end( $ids );
 		foreach ( $ids as $key => $id ) {
@@ -75,7 +73,7 @@ class Wpcli {
 	 * @param   boolean  $stdout    Optional. Clean stdout output.
 	 * @since   2.0.0
 	 */
-	private static function error( $code = 255, $stdout = false ) {
+	private function error( $code = 255, $stdout = false ) {
 		if ( \WP_CLI\Utils\isPiped() ) {
 			// phpcs:ignore
 			fwrite( STDOUT, '' );
@@ -83,11 +81,11 @@ class Wpcli {
 			exit( $code );
 		} elseif ( $stdout ) {
 			// phpcs:ignore
-			fwrite( STDERR, ucfirst( self::$exit_codes[ $code ] ) );
+			fwrite( STDERR, ucfirst( $this->exit_codes[ $code ] ) );
 			// phpcs:ignore
 			exit( $code );
 		} else {
-			\WP_CLI::error( self::$exit_codes[ $code ] );
+			\WP_CLI::error( $this->exit_codes[ $code ] );
 		}
 	}
 
@@ -99,7 +97,7 @@ class Wpcli {
 	 * @param   boolean  $stdout    Optional. Clean stdout output.
 	 * @since   2.0.0
 	 */
-	private static function warning( $msg, $result = '', $stdout = false ) {
+	private function warning( $msg, $result = '', $stdout = false ) {
 		if ( \WP_CLI\Utils\isPiped() || $stdout ) {
 			// phpcs:ignore
 			fwrite( STDOUT, $result );
@@ -116,7 +114,7 @@ class Wpcli {
 	 * @param   boolean  $stdout    Optional. Clean stdout output.
 	 * @since   2.0.0
 	 */
-	private static function success( $msg, $result = '', $stdout = false ) {
+	private function success( $msg, $result = '', $stdout = false ) {
 		if ( \WP_CLI\Utils\isPiped() || $stdout ) {
 			// phpcs:ignore
 			fwrite( STDOUT, $result );
@@ -133,7 +131,7 @@ class Wpcli {
 	 * @param   boolean  $stdout    Optional. Clean stdout output.
 	 * @since   2.0.0
 	 */
-	private static function line( $msg, $result = '', $stdout = false ) {
+	private function line( $msg, $result = '', $stdout = false ) {
 		if ( \WP_CLI\Utils\isPiped() || $stdout ) {
 			// phpcs:ignore
 			fwrite( STDOUT, $result );
@@ -149,7 +147,7 @@ class Wpcli {
 	 * @param   boolean  $stdout    Optional. Clean stdout output.
 	 * @since   2.0.0
 	 */
-	private static function log( $msg, $stdout = false ) {
+	private function log( $msg, $stdout = false ) {
 		if ( ! \WP_CLI\Utils\isPiped() && ! $stdout ) {
 			\WP_CLI::log( $msg );
 		}
@@ -162,7 +160,7 @@ class Wpcli {
 	 * @return  array The true parameters.
 	 * @since   2.0.0
 	 */
-	private static function get_params( $args ) {
+	private function get_params( $args ) {
 		$result = '';
 		if ( array_key_exists( 'settings', $args ) ) {
 			$result = \json_decode( $args['settings'], true );
@@ -184,7 +182,7 @@ class Wpcli {
 	 *     === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-opcache-manager/blob/master/WP-CLI.md ===
 	 *
 	 */
-	public static function status( $args, $assoc_args ) {
+	public function status( $args, $assoc_args ) {
 		\WP_CLI::line( sprintf( '%s is running.', Environment::plugin_version_text() ) );
 		$name = OPcache::name();
 		if ( '' === $name ) {
@@ -245,7 +243,7 @@ class Wpcli {
 	 *     === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-opcache-manager/blob/master/WP-CLI.md ===
 	 *
 	 */
-	public static function settings( $args, $assoc_args ) {
+	public function settings( $args, $assoc_args ) {
 		$stdout  = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
 		$action  = isset( $args[0] ) ? (string) $args[0] : '';
 		$setting = isset( $args[1] ) ? (string) $args[1] : '';
@@ -254,10 +252,10 @@ class Wpcli {
 				switch ( $setting ) {
 					case 'analytics':
 						Option::network_set( 'analytics', true );
-						self::success( 'analytics are now activated.', '', $stdout );
+						$this->success( 'analytics are now activated.', '', $stdout );
 						break;
 					default:
-						self::error( 1, $stdout );
+						$this->error( 1, $stdout );
 				}
 				break;
 			case 'disable':
@@ -265,14 +263,14 @@ class Wpcli {
 					case 'analytics':
 						\WP_CLI::confirm( 'Are you sure you want to deactivate analytics?', $assoc_args );
 						Option::network_set( 'analytics', false );
-						self::success( 'analytics are now deactivated.', '', $stdout );
+						$this->success( 'analytics are now deactivated.', '', $stdout );
 						break;
 					default:
-						self::error( 1, $stdout );
+						$this->error( 1, $stdout );
 				}
 				break;
 			default:
-				self::error( 2, $stdout );
+				$this->error( 2, $stdout );
 		}
 	}
 
@@ -304,10 +302,10 @@ class Wpcli {
 	 *    === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-opcache-manager/blob/master/WP-CLI.md ===
 	 *
 	 */
-	public static function analytics( $args, $assoc_args ) {
+	public function analytics( $args, $assoc_args ) {
 		$stdout = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
 		if ( ! Option::network_get( 'analytics' ) ) {
-			self::error( 3, $stdout );
+			$this->error( 3, $stdout );
 		}
 		$analytics = Analytics::get_status_kpi_collection();
 		$result    = [];
@@ -329,11 +327,11 @@ class Wpcli {
 		$format = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
 		if ( 'json' === $format ) {
 			$detail = wp_json_encode( $analytics );
-			self::line( $detail, $detail, $stdout );
+			$this->line( $detail, $detail, $stdout );
 		} elseif ( 'yaml' === $format ) {
 			unset( $analytics['assets'] );
 			$detail = Spyc::YAMLDump( $analytics, true, true, true );
-			self::line( $detail, $detail, $stdout );
+			$this->line( $detail, $detail, $stdout );
 		} else {
 			\WP_CLI\Utils\format_items( $assoc_args['format'], $result, [ 'kpi', 'description', 'value', 'ratio', 'variation' ] );
 		}
@@ -374,18 +372,18 @@ class Wpcli {
 	 *   === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-traffic/blob/master/WP-CLI.md ===
 	 *
 	 */
-	public static function exitcode( $args, $assoc_args ) {
+	public function exitcode( $args, $assoc_args ) {
 		$stdout = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
 		$format = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
 		$action = isset( $args[0] ) ? $args[0] : 'list';
 		$codes  = [];
-		foreach ( self::$exit_codes as $key => $msg ) {
+		foreach ( $this->exit_codes as $key => $msg ) {
 			$codes[ $key ] = [ 'code' => $key, 'meaning' => ucfirst( $msg ) ];
 		}
 		switch ( $action ) {
 			case 'list':
 				if ( 'ids' === $format ) {
-					self::write_ids( $codes );
+					$this->write_ids( $codes );
 				} else {
 					\WP_CLI\Utils\format_items( $format, $codes, [ 'code', 'meaning' ] );
 				}
@@ -401,7 +399,7 @@ class Wpcli {
 	 * @return  string  The output of the shortcode, ready to print.
 	 * @since 1.0.0
 	 */
-	public static function sc_get_helpfile( $attributes ) {
+	public function sc_get_helpfile( $attributes ) {
 		$md = new Markdown();
 		return $md->get_shortcode(  'WP-CLI.md', $attributes  );
 	}
@@ -411,8 +409,5 @@ class Wpcli {
 add_shortcode( 'opcm-wpcli', [ 'OPcacheManager\Plugin\Feature\Wpcli', 'sc_get_helpfile' ] );
 
 if ( defined( 'WP_CLI' ) && WP_CLI ) {
-	\WP_CLI::add_command( 'opcache status', [ Wpcli::class, 'status' ] );
-	\WP_CLI::add_command( 'opcache analytics', [ Wpcli::class, 'analytics' ] );
-	\WP_CLI::add_command( 'opcache settings', [ Wpcli::class, 'settings' ] );
-	\WP_CLI::add_command( 'opcache exitcode', [ Wpcli::class, 'exitcode' ] );
+	\WP_CLI::add_command( 'opcache', 'OPcacheManager\Plugin\Feature\Wpcli' );
 }